+++ lib/sysaid/ticket.rb
@@ -79,6 +79,18 @@ class SysAid::Ticket
     return ticket
   end
 
+  def self.find_by_query(query)
+    SysAid.ensure_logged_in
+
+    response = SysAid.call(:execute_select_query, message: sql_query(query))
+
+    if response.to_hash[:execute_select_query_response][:return]
+      return response.to_hash[:execute_select_query_response][:return]
+    end
+
+    return false
+  end
+
   # Loads the latest ticket information from the SysAid server
   def refresh
     SysAid.ensure_logged_in
@@ -134,7 +146,7 @@ class SysAid::Ticket
   #   => true
   def delete
     SysAid.ensure_logged_in
-    
+
     SysAid.call(:delete, message: to_xml(false))
 
     reset_all_attributes
@@ -261,5 +273,15 @@ class SysAid::Ticket
     self.user_manager = response[:user_manager]
     self.version = response[:version]
     self.workaround = response[:workaround]
+  end
+
+  def self.sql_query(query)
+    builder = Builder::XmlMarkup.new
+
+    builder.sessionId(SysAid.session_id)
+    xml = builder.apiSysObj('xsi:type' => "tns:apiServiceRequest")
+    xml = builder.condition(query)
+
+    xml
   end
 end